Adding attr_accessor for :passord in prod.
MG::Web does not have access to the actual member table in production
mode and we don't perform authentification ourselves. Because of this,
our table does not contain a password field. However, during login, it
is extremely practical for the model to still respond to the attribute,
as we then can utilize formtastic for the form.

This ensures that the model will respond to #passord and #passord= and
and corrects problems in production mode where the login form would
cause an error due to attribute not being present.