Ensure valid Member models in production.
When i was debugging why LogEntries would not save, it turned out
that LogEntries was validating its associated models (the Member model,
amongst others). In production, the 'passord' field would allways be
nil, as this is something that only exists in development.

By making this validation conditional on development, the validation
should return true in production. In the future, it might be worth to
unify the production and development Member models.